OpenCL comm + kernel 1x1 works

+++ OpenCL/device/lbp.cl
@@ -14,7 +14,7 @@ kernel void lbp(
     local float2* spoints,
 
     const int w, const int h,
-    const float radius,
+    const int radius,
     const int samples)
 {
     int xidx = get_local_id(0) + get_group_id(0) * get_local_size(0);
@@ -24,8 +24,8 @@ kernel void lbp(
     const int2 pos = { index / w, index % w };
 
     // Only continue if a distance 'radius' from the edge
-    if( pos.x < radius || pos.x >= w - radius || pos.y < radius || pos.y >= h - radius )
-        return;
+	if( pos.x < radius || pos.x >= w - radius || pos.y < radius || pos.y >= h - radius )
+		return;
 
     float a = 2.0f * M_PI / samples;
     for(int i = 0; i < samples; i++)
@@ -54,24 +54,24 @@ kernel void lbp(
         float current_value = in[index + ipos.y * w + ipos.x];
 
         // Check if interpolation is needed.
-        if( (abs_diff( fpos.x, ipos.x ) > 1e-6)
+		if( (abs_diff( fpos.x, ipos.x ) > 1e-6)
          || (abs_diff( fpos.y, ipos.y ) > 1e-6) )
         {
             int2 ff = (int2)floor(fpos);
             int2 cc = (int2)ceil(fpos);
 
             int idxV1 = index + w * ff.y + ff.x;
-            int idxV2 = index + w * ff.y + cc.x;
-            int idxV3 = index + w * cc.y + ff.x;
-            int idxV4 = index + w * cc.y + cc.x;
+			int idxV2 = index + w * ff.y + cc.x;
+			int idxV3 = index + w * cc.y + ff.x;
+			int idxV4 = index + w * cc.y + cc.x;
 
             // Calculate the interpolation weights.
-            float2 tt = fpos - ff;
+			float2 tt = fpos - ff;
 
-            float w1 = (1 - tt.x) * (1 - tt.y);
-            float w2 = tt.x * (1 - tt.y);
-            float w3 = (1 - tt.x) * tt.y;
-            float w4 = tt.x * tt.y;
+			float w1 = (1 - tt.x) * (1 - tt.y);
+			float w2 = tt.x * (1 - tt.y);
+			float w3 = (1 - tt.x) * tt.y;
+			float w4 = tt.x * tt.y;
 
             current_value   = in[idxV1] * w1
                             + in[idxV2] * w2
